Move back operators to its original files
